ckp on docker hosts
